requirements.txt
math
streamlit==1.13.0
numpy==1.21.2
opencv-python==4.6.0.66
Pillow==8.3.2
keras==2.10.0